[Test] Keep H3 VAE ops coverage self-contained
